pyproject.toml
# Note that in maintenance branches, all build dependencies should
# have an upper bound equal to the most recent already-released version
# of the dependency. This to prevent that a future backwards-incompatible
# release will break the source build of a SciPy release.
# Do accept micro (bug-fix) releases. So for example, if pybind11 2.4.3 is
# the most recent version on PyPI:
#
#     "pybind11>=2.4.3,<2.5.0",
#
# Upper bounds in release branches must have notes on why they are added.
# Distro packages can ignore upper bounds added only to prevent future
# breakage; if we add pins or bounds because of known problems then they need
# them too.

# Run tools/generate_requirements.txt when making changes to any dependencies

[build-system]
build-backend = 'mesonpy'
requires = [
    "meson-python>=0.15.0",
    "Cython>=3.0.8",        # when updating version, also update check in meson.build
    "pybind11>=2.13.2",     # when updating version, also update check in scipy/meson.build
    "pythran>=0.14.0",

    # MLX requirement for wheel builds
    "mlx>=0.20.0",
]

[project]
name = "scipy-mlx"
version = "1.17.0.dev0"
# TODO: add `license-files` once PEP 639 is accepted (see meson-python#88)
#       at that point, no longer include them in `py3.install_sources()`
license = { file = "LICENSE.txt" }
description = "Fundamental algorithms for scientific computing in Python"
maintainers = [
    { name = "SciPy Developers", email = "scipy-dev@python.org" },
]
# Note: Python and NumPy upper version bounds should be set correctly in
# release branches, see:
#     https://scipy.github.io/devdocs/dev/core-dev/index.html#version-ranges-for-numpy-and-other-dependencies
requires-python = ">=3.11"  # keep in sync with `min_python_version` in meson.build
dependencies = [
    "mlx>=0.20.0",
] # MLX as the array backend
readme = "README.rst"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: C",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-cov",
    "pytest-timeout",
    "pytest-xdist",
    "asv",
    "mpmath",
    "gmpy2",
    "threadpoolctl",
    "scikit-umfpack",
    "pooch",
    "hypothesis>=6.30",
    "array-api-strict>=2.3.1",
    "Cython",
    "meson",
    'ninja; sys_platform != "emscripten"',
]
doc = [
    "sphinx>=5.0.0,<8.2.0",
    "intersphinx_registry",
    "pydata-sphinx-theme>=0.15.2",
    "sphinx-copybutton",
    "sphinx-design>=0.4.0",
    "matplotlib>=3.5",
    "numpydoc",
    "jupytext",
    "myst-nb>=1.2.0",
    "pooch",
    "jupyterlite-sphinx>=0.19.1",
    "jupyterlite-pyodide-kernel",
    "linkify-it-py",
    "tabulate"
]

# click is a dependency of spin, and not a direct SciPy dependency,
# but click 8.3.0 is apparently broken, so use an older version.
dev = [
    "click<8.3.0",
    "spin",
    "mypy==1.10.0",
    "typing_extensions",
    "types-psutil",
    "pycodestyle",
    "ruff>=0.12.0",
    "cython-lint>=0.12.2",
]

[project.urls]
homepage = "https://scipy.org/"
documentation = "https://docs.scipy.org/doc/scipy/"
source = "https://github.com/scipy/scipy"
download = "https://github.com/scipy/scipy/releases"
tracker = "https://github.com/scipy/scipy/issues"

[tool.spin]
package = 'scipy_mlx'

[tool.meson-python.args]
install = ['--skip-subprojects']

[tool.cibuildwheel]
skip = ["*_i686", "*_ppc64le", "*_s390x", "*_universal2"]
# We're only testing with essential test dependencies, not optional ones.
# Some of those require binary wheels (often missing for some platforms),
# or they slow down the test suite runs too much or simply aren't necessary.
test-requires = [
    "pytest",
    "pytest-xdist",
    "threadpoolctl",
    "pooch",
    "hypothesis",
]
test-command = "bash {project}/tools/wheels/cibw_test_command.sh {project}"

[tool.cibuildwheel.linux]
manylinux-x86_64-image = "manylinux_2_28"
manylinux-aarch64-image = "manylinux_2_28"
musllinux-x86_64-image = "musllinux_1_2"
musllinux-aarch64-image = "musllinux_1_2"
before-build = "bash {project}/tools/wheels/cibw_before_build_linux.sh {project}"

[tool.cibuildwheel.linux.environment]
# /project will be the $PWD equivalent inside the docker used to build the wheel
PKG_CONFIG_PATH = "/project/"

[tool.cibuildwheel.macos]
before-build = "bash {project}/tools/wheels/cibw_before_build_macos.sh {project}"

[tool.cibuildwheel.macos.environment]
PKG_CONFIG_PATH = "{project}"

[tool.cibuildwheel.windows]
before-build = "bash {project}/tools/wheels/cibw_before_build_win.sh {project}"
repair-wheel-command = "bash ./tools/wheels/repair_windows.sh {wheel} {dest_dir}"

[tool.cibuildwheel.windows.environment]
# This does not work because pkg-config does not like backslashes,
PKG_CONFIG_PATH = "{project}"
# do this instead (which will override this setting)
# set CIBW_ENVIRONMENT_WINDOWS=PKG_CONFIG_PATH=PWD.replace('\\', '/')

[tool.spin.commands]
"Build" = [
  ".spin/cmds.py:build",
  ".spin/cmds.py:test",
  ".spin/cmds.py:mypy",
  ".spin/cmds.py:lint",
  "spin.cmds.pip.install"
]
"Environments" = [
  "spin.cmds.meson.run",
  ".spin/cmds.py:python",
  ".spin/cmds.py:ipython",
  ".spin/cmds.py:shell",
  "spin.cmds.meson.gdb",
  "spin.cmds.meson.lldb"
]
"Documentation" = [
  ".spin/cmds.py:docs",
  ".spin/cmds.py:smoke_docs",
  ".spin/cmds.py:refguide_check",
  ".spin/cmds.py:smoke_tutorials"
]
"Release" = [
  ".spin/cmds.py:notes",
  ".spin/cmds.py:authors"
]
"Metrics" = [".spin/cmds.py:bench"]
